GUI-258. Move account and currency types anf tcomb validators to entities.
