+++ front/src/app/api/mock.ts
@@ -1,6 +1,10 @@
 import * as t from './types';
-import { ICurrencyInfo } from 'app/entities/currency';
-import { IAccountInfo } from 'app/entities/account';
+import { ICurrencyInfo, ICurrencyBalanceMap } from 'app/entities/currency';
+import {
+    IAccountInfo,
+    EnumProfileStatus,
+    IProfileBrief,
+} from 'app/entities/account';
 
 const VASYA_ADDR = '0x88057f14236687831e1fd205e8efb9e45166fe72';
 const PETYA_ADDR = '0xfd0c80ba15cbf19770319e5e76ae05012314608f';
@@ -91,7 +95,7 @@ export const accountListResult: t.IResult<IAccountInfo[]> = {
             marketUsdBalance: '12345678',
             marketBalance: '0987654321',
             currencyBalanceMap: (currencyListResult.data as ICurrencyInfo[]).reduce(
-                (acc: t.ICurrencyBalanceMap, currency) => {
+                (acc: ICurrencyBalanceMap, currency) => {
                     acc[currency.address] = String(Math.random() * 100).substr(
                         0,
                         15,
@@ -109,7 +113,7 @@ export const accountListResult: t.IResult<IAccountInfo[]> = {
             marketUsdBalance: '12345678',
             marketBalance: '0987654321',
             currencyBalanceMap: (currencyListResult.data as ICurrencyInfo[]).reduce(
-                (acc: t.ICurrencyBalanceMap, currency) => {
+                (acc: ICurrencyBalanceMap, currency) => {
                     acc[currency.address] = String(Math.random() * 100).substr(
                         0,
                         15,
@@ -157,7 +161,7 @@ export function send(
     return result;
 }
 
-export const profileListResult: t.IResult<t.IListResult<t.IProfileBrief>> = {
+export const profileListResult: t.IResult<t.IListResult<IProfileBrief>> = {
     data: {
         records: [
             {
@@ -168,7 +172,7 @@ export const profileListResult: t.IResult<t.IListResult<t.IProfileBrief>> = {
                 deals: 98,
                 country: 'ru',
                 logoUrl: '',
-                status: t.EnumProfileStatus.anon,
+                status: EnumProfileStatus.anon,
             },
             {
                 name: '',
@@ -178,7 +182,7 @@ export const profileListResult: t.IResult<t.IListResult<t.IProfileBrief>> = {
                 deals: 98,
                 country: 'ru',
                 logoUrl: '',
-                status: t.EnumProfileStatus.anon,
+                status: EnumProfileStatus.anon,
             },
             {
                 name: 'Ololan Ololoshevich',
@@ -188,7 +192,7 @@ export const profileListResult: t.IResult<t.IListResult<t.IProfileBrief>> = {
                 deals: 98,
                 country: 'gb',
                 logoUrl: '',
-                status: t.EnumProfileStatus.anon,
+                status: EnumProfileStatus.anon,
             },
             {
                 name: 'BG',
@@ -198,7 +202,7 @@ export const profileListResult: t.IResult<t.IListResult<t.IProfileBrief>> = {
                 deals: 98,
                 country: 'us',
                 logoUrl: '',
-                status: t.EnumProfileStatus.anon,
+                status: EnumProfileStatus.anon,
             },
         ],
         total: 5,
