+++ front/src/app/stores/my-profiles.ts
@@ -4,10 +4,10 @@ import { OnlineStore, IOnlineStoreServices } from './online-store';
 import { observable, computed, action, reaction, IReactionPublic } from 'mobx';
 import { asyncAction } from 'mobx-utils';
 import { IAccountItemView } from 'app/stores/types';
-import Api, { IProfileFull, IMarketStats } from 'app/api';
+import Api, { IMarketStats } from 'app/api';
 import { IValidation } from 'app/localization/types';
 import { RootStore } from 'app/stores';
-import { IAccountInfo } from 'app/entities/account';
+import { IAccountInfo, IProfileFull } from 'app/entities/account';
 const { pending, catchErrors } = OnlineStore;
 import { createBigNumber, ZERO, BN } from '../utils/create-big-number';
 import { ICurrencyInfo } from 'app/entities/currency';
