+++ front/src/worker/api/dwh.ts
@@ -2,12 +2,18 @@ import * as t from './types';
 // import { IAttribute } from '../../app/api/types';
 import * as mapKeys from 'lodash/fp/mapKeys';
 import * as pick from 'lodash/fp/pick';
-import { TypeEthereumAddress } from '../../app/api/runtime-types';
 import * as tcomb from 'tcomb';
 import { BN } from 'bn.js';
-import { EnumProfileStatus, IBenchmarkMap } from '../../app/api/types';
+import { IBenchmarkMap } from '../../app/api/types';
 import * as moment from 'moment';
 import * as get from 'lodash/fp/get';
+import {
+    EnumProfileStatus,
+    ICertificate,
+    IProfileBrief,
+    IProfileFull,
+} from 'app/entities/account';
+import { TypeEthereumAddress } from 'app/entities/currency';
 
 interface IDictionary<T> {
     [index: string]: keyof T;
@@ -58,7 +64,7 @@ export class DWH {
         [8, 'gpuRamSize', MB_SIZE],
     ];
 
-    public static readonly mapProfile: IDictionary<t.IProfileBrief> = {
+    public static readonly mapProfile: IDictionary<IProfileBrief> = {
         UserID: 'address',
         IdentityLevel: 'status',
         Name: 'name',
@@ -71,10 +77,10 @@ export class DWH {
         (Object as any).values(DWH.mapProfile),
     );
 
-    public static readonly defaultProfile: t.IProfileBrief = {
+    public static readonly defaultProfile: IProfileBrief = {
         name: '',
         address: '0x0',
-        status: t.EnumProfileStatus.anonimest,
+        status: EnumProfileStatus.anonimest,
         sellOrders: 0,
         buyOrders: 0,
         deals: 0,
@@ -97,17 +103,17 @@ export class DWH {
     };
 
     public static kycAttributesToStatus = {
-        [t.EnumAttributes.Kyc2]: t.EnumProfileStatus.reg,
-        [t.EnumAttributes.Kyc3]: t.EnumProfileStatus.ident,
-        [t.EnumAttributes.Kyc4]: t.EnumProfileStatus.ident,
+        [t.EnumAttributes.Kyc2]: EnumProfileStatus.reg,
+        [t.EnumAttributes.Kyc3]: EnumProfileStatus.ident,
+        [t.EnumAttributes.Kyc4]: EnumProfileStatus.ident,
     };
 
-    private processProfile(item: any): t.IProfileBrief {
+    private processProfile(item: any): IProfileBrief {
         const renamed = DWH.renameProfileKeys(item);
         const picked = DWH.pickProfileKeys(renamed);
         const result = { ...DWH.defaultProfile, ...picked };
 
-        return result as t.IProfileBrief;
+        return result as IProfileBrief;
     }
 
     public getProfiles = async ({
@@ -116,7 +122,7 @@ export class DWH {
         offset,
         sortBy,
         sortDesc,
-    }: t.IListQuery): Promise<t.IListResult<t.IProfileBrief>> => {
+    }: t.IListQuery): Promise<t.IListResult<IProfileBrief>> => {
         tcomb.maybe(tcomb.String)(filter);
         tcomb.Number(limit);
         tcomb.Number(offset);
@@ -177,9 +183,7 @@ export class DWH {
         };
     };
 
-    public getProfileFull = async ({
-        address,
-    }: any): Promise<t.IProfileFull> => {
+    public getProfileFull = async ({ address }: any): Promise<IProfileFull> => {
         TypeEthereumAddress(address);
 
         const res = await this.fetchData('GetProfileInfo', { Id: address });
@@ -542,7 +546,7 @@ export class DWH {
         return attrMap;
     }
 
-    protected getKycCertificates(certificates: any[]): t.ICertificate[] {
+    protected getKycCertificates(certificates: any[]): ICertificate[] {
         return certificates
             .filter(x => x.attribute in DWH.kycAttributesToStatus)
             .map(x => ({
